pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.59.0",
    "generativeai>=0.0.1",
    "google-genai>=1.27.0",
    "google-generativeai>=0.8.5",
    "mcp>=1.12.2",
    "python-dotenv>=1.1.1",
]
